\name{tulln}
\alias{tulln}
\alias{tu1}
\alias{vs50}
\title{Data on Chlorid concentrations in the Suedliche Tullnerfeld}

\description{
The \code{tulln} dataset contains a variogram object and a matrix
with the coordinates of the vertices of a polygonal region.

}

\format{
  This dataset consists of a variogram object and a matrix.
  }

\details{
The variogram model used is a spherical model.
}

\source{
 Glatzer, E and Mueller, W.G. (2002).
}

\references{ E. Glatzer, W.G. Mueller, 2004. Residual diagnostics for
variogram fitting. Computers and Geosciences 30, pp 859-866. }

\usage{
data(tulln)
}
\keyword{datasets}
